# infrastructure/swagger/api/gen/proto.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/v1/authentication.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/SoundService/DiarizateFile": {
      "get": {
        "summary": "Send an audio file for transcription with audio diarization",
        "operationId": "SoundService_DiarizateFile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/sound_transferSpeakerAndLineResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "soundData",
            "description": "Selected file's raw bytes.",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "byte"
          },
          {
            "name": "sourceLanguage",
            "description": "Selected file's language.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "SoundService"
        ]
      }
    },
    "/SoundService/TestConnection": {
      "get": {
        "summary": "Test connection to the gRPC server",
        "operationId": "SoundService_TestConnection",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/sound_transferTextMessage"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "SoundService"
        ]
      }
    },
    "/SoundService/TranscribeFile": {
      "get": {
        "summary": "Send an audio file for transcription",
        "operationId": "SoundService_TranscribeFile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/sound_transferSoundResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "soundData",
            "description": "Selected file's raw bytes.",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "byte"
          },
          {
            "name": "sourceLanguage",
            "description": "Selected file's language.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "SoundService"
        ]
      }
    },
...
    "/SoundService/TranslateFile": {
      "get": {
        "summary": "Send an audio file for translation",
        "operationId": "SoundService_TranslateFile",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/sound_transferSoundResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of sound_transferSoundResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "soundData",
            "description": "Selected file's raw bytes.",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "byte"
          },
          {
            "name": "sourceLanguage",
            "description": "Selected file's language.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "translationLanguage",
            "description": "Desired translation language.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "SoundService"
        ]
      }
    }
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "sound_transferSoundResponse": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string",
          "title": "Transcription of the audio file"
        },
        "detectedLanguage": {
          "type": "string",
          "title": "Detected language by speech-to-text model"
        }
      },
      "title": "sound_service.SoundResponse"
    },
...
    "sound_transferSpeakerAndLineResponse": {
      "type": "object",
      "properties": {
        "text": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Array of transcriptions of the audio file, order preserved"
        },
        "speakerName": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Array of recognised speakers in the audio file, order preserved"
        },
        "detectedLanguage": {
          "type": "string",
          "title": "Detected language by speech-to-text model"
        }
      },
      "title": "sound_service.SpeakerAndLineResponse"
    },
    "sound_transferTextMessage": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string"
        }
      },
      "title": "sound_service.TextMessage"
    }
  }
}